moved util functions in util file
